requirements.txt
openai==0.27.4
datasets
jsonlines
colorama
python-dotenv
tiktoken